Fixed code analysis tools warnings
Fixed the most important code analysis tools warnings for us to start working on the project. There are still some not fixed, but where left unfixed because they are very likely to be solved by some future already known changes in project structure. It was pointless to fix them now, knowing those parts of the project will be changed in the future fixing the warnings.